resurrected breadcrumb styles and refactored breadcrumb comp vs breadcrumb region
