+++ components/components/breadcrumb/breadcrumb.twig
@@ -1,0 +1,25 @@
+{% set classes = [
+  'c-breadcrumb',
+  modifier_classes ? modifier_classes : '',
+]|join(' ')|trim %}
+
+<nav {{ add_attributes({
+  'aria-label': label,
+  'class': classes,
+}) }}>
+  <div class="l-constrain">
+    <ol class="c-breadcrumb__list">
+      {% for item in breadcrumb %}
+        <li class="c-breadcrumb__item">
+          {% if item.url %}
+            <a class="c-breadcrumb__link" href="{{ item.url }}">{{ item.text }}</a>
+          {% else %}
+            <span aria-current="page">
+              {{ item.text }}
+            </span>
+          {% endif %}
+        </li>
+      {% endfor %}
+    </ol>
+  </div>
+</nav>
