+++ components/components/breadcrumb/breadcrumb.yml
@@ -1,0 +1,15 @@
+---
+modifier_classes: ''
+label: 'Breadcrumb'
+breadcrumb:
+  -
+    url: '#0'
+    text: 'Home'
+  -
+    url: '#0'
+    text: 'Level 1'
+  -
+    url: '#0'
+    text: 'Level 2'
+  -
+    text: 'Current item'
