+++ components/layouts/breadcrumb-wrapper/breadcrumb-wrapper.component.yml
@@ -1,5 +1,5 @@
 $schema: https://git.drupalcode.org/project/drupal/-/raw/HEAD/core/assets/schemas/v1/metadata.schema.json
-name: Breadcrumb
+name: Breadcrumb Wrapper
 props:
   type: object
   properties:
