+++ components/layouts/breadcrumb-wrapper/breadcrumb-wrapper.twig
@@ -1,5 +1,5 @@
 {% set classes = [
-  'l-breadcrumb',
+  'l-breadcrumb-wrapper',
   modifier_classes ? modifier_classes : '',
 ]|join(' ')|trim %}
