--- components/layouts/breadcrumb/breadcrumb.source.scss
@@ -1,3 +1,0 @@
-// Layout: Breadcrumb
-
-// .l-breadcrumb {}
