+++ templates/layout/region--breadcrumb.html.twig
@@ -5,7 +5,7 @@
  */
 #}
 {% if content %}
-  {% embed 'gesso:breadcrumb' with {
+  {% embed 'gesso:breadcrumb-wrapper' with {
     'attributes': attributes,
     'has_constrain': false,
   } %}
